update filters and models

+++ src/Waste2MealsClient/Models/Requests/UpdateBatchDefinitionRequest.cs
@@ -6,8 +6,8 @@ public record UpdateBatchDefinitionRequest(
     string Description,
     string Tag,
     string ImageUrl, 
-    double OriginalPrice, 
-    double DiscountPrice,
+    decimal OriginalPrice, 
+    decimal DiscountPrice,
     TimeSpan PickupStartTime,
     TimeSpan PickupEndTime
 );
